Move unboxing helpers to CastHelpers class
